ffeat: class unit testing

+++ usecase/class/delete.go
@@ -2,6 +2,7 @@ package class
 
 import (
 	"context"
+	"time"
 
 	"github.com/teq-quocbang/course-register/util/myerror"
 )
@@ -11,9 +12,17 @@ func (u *UseCase) Delete(ctx context.Context, id string) error {
 		return myerror.ErrClassInvalidParam("id")
 	}
 
-	_, err := u.Class.GetByID(ctx, id)
+	class, err := u.Class.GetByID(ctx, id)
 	if err != nil {
 		return myerror.ErrClassGet(err)
+	}
+
+	if class.StartTime.Before(time.Now()) && class.EndTime.After(time.Now()) {
+		return myerror.ErrClassInvalidParam("class is going on")
+	}
+
+	if class.EndTime.Before(time.Now()) {
+		return myerror.ErrClassInvalidParam("class ended")
 	}
 
 	err = u.Class.Delete(ctx, id)
